pkgbuild-package {pkgbuild}	R Documentation

pkgbuild: Find Tools Needed to Build R Packages

Description

Provides functions used to build R packages. Locates compilers needed to build R packages on various platforms and ensures the PATH is configured appropriately so R can use them.

Author(s)

Maintainer: Gábor Csárdi csardi.gabor@gmail.com

Authors:

Hadley Wickham

Jim Hester

Other contributors:

Posit Software, PBC [copyright holder, funder]

See Also

Useful links:

https://github.com/r-lib/pkgbuild

https://pkgbuild.r-lib.org

Report bugs at https://github.com/r-lib/pkgbuild/issues

[Package pkgbuild version 1.4.4 Index]