has_build_tools {pkgbuild}	R Documentation

Are build tools are available?

Description

has_build_tools returns a logical, check_build_tools throws an error. with_build_tools checks that build tools are available, then runs code in an correctly staged environment. If run interactively from RStudio, and the build tools are not available these functions will trigger an automated install.

Usage

has_build_tools(debug = FALSE)

check_build_tools(debug = FALSE)

with_build_tools(code, debug = FALSE, required = TRUE)

Arguments

debug
If TRUE, will print out extra information useful for debugging. If FALSE, it will use result cached from a previous run.

code
Code to rerun in environment where build tools are guaranteed to exist.

required
If TRUE, and build tools are not available, will throw an error. Otherwise will attempt to run code without them.

Examples

has_build_tools()
check_build_tools()

[Package pkgbuild version 0.0.0.9000 Index]