has_compiler {pkgbuild}	R Documentation

Is a compiler available?

Description

has_devel returns TRUE or FALSE. check_devel throws an error if you don't have developer tools installed. Implementation based on a suggestion by Simon Urbanek. End-users (particularly those on Windows) should generally run check_build_tools() rather than check_compiler().

Usage

has_compiler(debug = FALSE)

check_compiler(debug = FALSE)

Arguments

debug
If TRUE, will print out extra information useful for debugging. If FALSE, it will use result cached from a previous run.

See Also

check_build_tools()

Examples

has_compiler()
check_compiler()

with_build_tools(has_compiler())

[Package pkgbuild version 1.3.1 Index]