needs_compile {pkgbuild}	R Documentation

Does the package need recompiling? (i.e. is there a source or header file newer than the dll)

Description

Does the package need recompiling? (i.e. is there a source or header file newer than the dll)

Usage

needs_compile(path = ".")

Arguments

path
Path to a package, or within a package.

[Package pkgbuild version 1.3.1 Index]