pkg_links_to_rcpp {pkgbuild}	R Documentation

Test if a package path is linking to Rcpp or cpp11

Description

Test if a package path is linking to Rcpp or cpp11

Usage

pkg_links_to_rcpp(path)

pkg_links_to_cpp11(path)

Arguments

path
Path to a package, or within a package.

[Package pkgbuild version 1.3.1 Index]