rtools_needed {pkgbuild}	R Documentation

Retrieve a text string with the rtools version needed

Description

Retrieve a text string with the rtools version needed

Usage

rtools_needed(r_version = getRversion())

[Package pkgbuild version 1.3.1 Index]